requirements.txt
fastapi
streamlit
pandas
numpy
scikit-learn
xgboost
lightgbm
joblib
uvicorn[standard]
python-dotenv
shap
pyyaml
matplotlib
pytest
plotly